Modularise `bindJsEvaluator`
Issue: #298
